chore: extension 코드 한곳에 모음
