+++ walkmong/walkmong/Extension/UIView+.swift
@@ -70,4 +70,5 @@ extension UIView {
         view.layer.masksToBounds = true
         return view
     }
+    
 }
